requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1
mysqlclient==2.2.7
pymysql==1.0.3
